move product-view to components page
