+++ packages/template-retail-react-app/app/components/product-view/index.jsx
@@ -15,16 +15,16 @@ import {useDerivedProduct} from '../../hooks'
 import {useAddToCartModalContext} from '../../hooks/use-add-to-cart-modal'
 
 // project components
-import SwatchGroup from '../../components/swatch-group'
-import Swatch from '../../components/swatch-group/swatch'
-import ImageGallery from '../../components/image-gallery'
-import Breadcrumb from '../../components/breadcrumb'
-import Link from '../../components/link'
+import SwatchGroup from '../swatch-group'
+import Swatch from '../swatch-group/swatch'
+import ImageGallery from '../image-gallery'
+import Breadcrumb from '../breadcrumb'
+import Link from '../link'
 import withRegistration from '../../hoc/with-registration'
 import {useCurrency} from '../../hooks'
-import {Skeleton as ImageGallerySkeleton} from '../../components/image-gallery'
-import {HideOnDesktop, HideOnMobile} from '../../components/responsive'
-import QuantityPicker from '../../components/quantity-picker'
+import {Skeleton as ImageGallerySkeleton} from '../image-gallery'
+import {HideOnDesktop, HideOnMobile} from '../responsive'
+import QuantityPicker from '../quantity-picker'
 import {useToast} from '../../hooks/use-toast'
 import {API_ERROR_MESSAGE} from '../../constants'
