+++ packages/template-retail-react-app/app/components/product-view/index.test.js
